agent_api/requirements.txt
# Core dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
httpx>=0.26.0

# Pydantic AI
pydantic-ai>=0.0.9
pydantic-ai-slim[mcp]>=0.4.0,<1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0,<3.0.0

# Database
supabase>=2.3.0
vecs>=0.4.0

# OpenAI (for embeddings and title generation)
openai>=1.12.0

# Mem0 for memory management
mem0ai>=0.1.0

# Search API clients
brave-search-python-client>=0.4.0,<1.0.0

# Additional FastAPI dependencies
python-multipart>=0.0.9

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0,<24.0.0
mypy>=1.0.0,<2.0.0

# Optional: Logging and monitoring
logfire>=0.1.0,<1.0.0
devtools>=0.12.0,<1.0.0